Code cleanup and adding search functionality
This commit adds a `find` command to find a server with certain properties. Called the same way as `add` but only include the properties you want to match.

Switched to using `nodemon` so making changes to the code automatically restart the node list server (remember that you will need to repopulate the list of multiplayer servers).

The existing code was inconsistent about whether to prefix properties with "server" or not. E.g., it uses "ip" and "serverPort." These inconsistently named properties were making it difficult to check for property matches in find, so I named them consistently by getting rid of the redundant server, so the properties are now "ip" and "port." Note that you will need to make similar changes in your postman queries.